Decompile EntityTeleport in RIC and DRA
