+++ config/splat.us.ric.yaml
@@ -51,6 +51,8 @@ segments:
       - [0x1A918, .rodata, 24788]
       - [0x1A9D0, .rodata, 26C84]
       - [0x1A9E8, rodata]
+      - [0x1AA28, .rodata, 2A060]
+      - [0x1AAA0, rodata]
       - [0x1AB04, rodata]
       - [0x1AB68, .rodata, 32324]
       - [0x1AC50, rodata]
@@ -62,6 +64,7 @@ segments:
       - [0x21250, c, 21250]
       - [0x24788, c, 24788]
       - [0x26C84, c, 26C84]
+      - [0x2A060, c, 2A060]
       - [0x32324, c, 32324]
       - [0x37B64, sbss]
   - [0x39A58]
