+++ src/dra/7E4BC.c
@@ -906,6 +906,7 @@ void func_80123F78(Entity* entity) {
     }
 }
 
+// Corresponding RIC function is func_80165DD8
 void func_80124164(
     POLY_GT4* poly, s32 colorIntensity, s32 y, s32 radius, bool arg4) {
     s16 top = y - radius;
@@ -1007,258 +1008,4 @@ void func_80124164(
     }
     poly->b2 = (u32)colorChannel >> 8;
     poly->b0 = (u32)colorChannel >> 8;
-}
-
-// teleport effect like when using library card (ID 0x42)
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntityTeleport);
-
-void func_80124A8C(Entity* entity) {
-#ifdef PSY_Q_3_5
-    if (PLAYER.step != 0 ||
-        PLAYER.step_s != 4) { // REAL: Matches with PSY-Q 3.5
-#else
-    if (*(s32*)&PLAYER.step != 0x40000) { // !FAKE: Probably ASPSX
-#endif
-        DestroyEntity(entity);
-        return;
-    }
-
-    switch (entity->step) {
-    case 0:
-        entity->animSet = ANIMSET_DRA(17);
-        entity->velocityY = FIX(-0.375);
-        SetSpeedX(0x4000);
-        entity->unk5A = 0x50;
-        entity->palette = 0x819F;
-        entity->unk4C = &D_800AE294;
-        entity->flags = FLAG_UNK_100000;
-        entity->facingLeft = 0;
-        entity->posY.i.hi -= 16;
-        entity->posX.val += entity->velocityX << 5;
-        entity->step++;
-        break;
-
-    case 1:
-        entity->posX.val += entity->velocityX;
-        entity->posY.val += entity->velocityY;
-        if (entity->animFrameDuration < 0) {
-            DestroyEntity(entity);
-        }
-        break;
-    }
-}
-
-// dagger thrown when using subweapon
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySubwpnThrownDagger);
-
-// axe thrown when using subweapon
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySubwpnThrownAxe);
-
-s32 func_80125A30(s32 baseY, s32 baseX) {
-    s16 x;
-    s16 y;
-    Collider res1;
-    Collider res2;
-    s16 colRes1;
-    s16 colRes2;
-
-    x = baseX + g_CurrentEntity->posX.i.hi;
-    y = baseY + g_CurrentEntity->posY.i.hi;
-
-    CheckCollision(x, y, &res1, 0);
-    colRes1 = res1.effects & 0xF801;
-    CheckCollision(x, (s16)(y - 1 + res1.unk18), &res2, 0);
-    y = baseY + (g_CurrentEntity->posY.i.hi + res1.unk18);
-
-    if ((colRes1 & 0x8801) == 1 || (colRes1 & 0x8801) == 0x0801) {
-        colRes2 = res2.effects & 0xF001;
-        if (!((s16)res2.effects & 1)) {
-            g_CurrentEntity->posY.i.hi = y;
-            return 1;
-        }
-        if ((res2.effects & 0x8001) == 0x8001) {
-            g_CurrentEntity->posY.i.hi = y + (s16)(res2.unk18 - 1);
-            return colRes2;
-        }
-    } else if ((colRes1 & 0x8001) == 0x8001) {
-        g_CurrentEntity->posY.i.hi = y;
-        return colRes1 & 0xF001;
-    }
-    return 0;
-}
-
-s32 func_80125B6C(s16 arg0, s16 arg1) {
-    Collider collider;
-    s16 var_a1;
-
-    if (g_CurrentEntity->velocityX == 0) {
-        return 0;
-    }
-
-    CheckCollision(g_CurrentEntity->posX.i.hi + arg1,
-                   g_CurrentEntity->posY.i.hi + arg0, &collider, 0);
-    if (g_CurrentEntity->velocityX > 0) {
-        var_a1 = collider.unk14;
-    } else {
-        var_a1 = collider.unk1C;
-    }
-
-    if (collider.effects & EFFECT_UNK_0002) {
-        g_CurrentEntity->posX.i.lo = 0;
-        g_CurrentEntity->posX.i.hi += var_a1;
-        return 2;
-    }
-
-    return 0;
-}
-
-void EntityHolyWater(Entity* entity) {
-    s16 temp3;
-    s32 temp;
-    s32 temp2;
-
-    switch (entity->step) {
-    case 0:
-        entity->flags = FLAG_UNK_08000000;
-        entity->animSet = ANIMSET_DRA(9);
-        entity->animCurFrame = 0x1D;
-        entity->zPriority = PLAYER.zPriority - 2;
-        entity->posY.i.hi += 8;
-        SetSpeedX(0x14000);
-        entity->velocityY = FIX(-3.125);
-        func_8011A290(entity);
-        entity->hitboxHeight = 4;
-        entity->hitboxWidth = 4;
-        g_Player.D_80072F00[10] = 4;
-        entity->step++;
-        break;
-
-    case 1:
-        entity->posY.val += entity->velocityY;
-        if (entity->velocityY <= 0x3FFFF) {
-            entity->velocityY += FIX(0.21875);
-        }
-
-        temp = func_80125A30(0, 0);
-        entity->posX.val += entity->velocityX;
-
-        if (entity->velocityX < 0) {
-            temp3 = -4;
-        } else {
-            temp3 = 4;
-        }
-        temp |= func_80125B6C(-7, temp3);
-
-        if (temp & 2) {
-            temp = 1;
-        } else {
-            temp2 = 1;
-        }
-
-        temp2 = temp & 1;
-        if (temp2 != 0) {
-            PlaySfx(0x69A);
-            CreateEntFactoryFromEntity(entity, FACTORY(0, 59), 0);
-            entity->ext.generic.unk7C.s = 0x10;
-            entity->animSet = ANIMSET_DRA(0);
-            entity->step = 2;
-        }
-        break;
-
-    case 2:
-        if (!(entity->ext.generic.unk7C.s & 3)) {
-            CreateEntFactoryFromEntity(entity, FACTORY(D_8013841C << 8, 28),
-                                       entity->ext.generic.unkB2 << 9);
-            D_8013841C++;
-        }
-        entity->ext.generic.unk7C.s--;
-        if (entity->ext.generic.unk7C.s == 0) {
-            entity->ext.generic.unk7C.s = 4;
-            entity->step++;
-        }
-        break;
-
-    case 3:
-        entity->ext.generic.unk7C.s--;
-        if (entity->ext.generic.unk7C.s == 0) {
-            DestroyEntity(entity);
-        }
-        break;
-    }
-}
-
-// glass breaking effect for holy water
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntityHolyWaterBreakGlass);
-
-// green flame when holy water explodes
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntityHolyWaterFlame);
-
-// cross subweapon crash (full effect with all parts)
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySubwpnCrashCross);
-
-// rising blue particles from cross crash
-void EntitySubwpnCrashCrossParticles(Entity* self) {
-    Primitive* prim;
-    u16 rand63;
-
-    if (self->step == 0) {
-        self->primIndex = AllocPrimitives(PRIM_GT4, 64);
-        if (self->primIndex != -1) {
-            self->flags = FLAG_UNK_04000000 | FLAG_HAS_PRIMS;
-            // entity lives for 192 frames
-            self->ext.generic.unk7C.s = 192;
-            self->step++;
-            return;
-        }
-        DestroyEntity(self);
-        return;
-    }
-    // This is some kind of time to live, since it decrements and if 0 gets
-    // destroyed.
-    if (--self->ext.generic.unk7C.s == 0) {
-        DestroyEntity(self);
-        return;
-    }
-    // On every third frame, as long as we have over 9 frames left alive
-    if ((self->ext.generic.unk7C.s >= 9) && !(self->ext.generic.unk7C.s & 3)) {
-        // iterate through primtives until we find one where r0 == 0, and set to
-        // 1
-        for (prim = &g_PrimBuf[self->primIndex]; prim != NULL;
-             prim = prim->next) {
-            if (prim->r0 == 0) {
-                prim->r0 = 1;
-                prim->r1 = 0;
-                break;
-            }
-        }
-    }
-
-    for (prim = &g_PrimBuf[self->primIndex]; prim != NULL; prim = prim->next) {
-        // for any of those prims with nonzero r0 values,
-        if (prim->r0 != 0) {
-            // r1 acts as a flag to show whether this has happened.
-            if (prim->r1 == 0) {
-                rand63 = rand() & 0x3F; // random integer 0-63
-                prim->g0 = (rand() % 237) + 9;
-                prim->g1 = -0x10 - (rand() & 0x20);
-                prim->clut = 0x1B0;
-                prim->tpage = 0x1A;
-                prim->b0 = 0;
-                prim->b1 = 0;
-                prim->priority = (rand63 + PLAYER.zPriority) - 0x20;
-                prim->blendMode = 0;
-                prim->g3 = (rand63 >> 2) + 4; // rand15 + 4 means 4 to 19
-                prim->r1++;
-            } else {
-                prim->g1 -= prim->g3;
-                if (((u8)prim->b0 >= 6U) || ((u8)prim->g1 < 0x18U)) {
-                    prim->blendMode = BLEND_VISIBLE;
-                    prim->r0 = 0;
-                }
-            }
-            if (prim->r0 != 0) {
-                func_80119E78(prim, prim->g0, prim->g1);
-            }
-        }
-    }
 }
